Added tests for sparse features, moved feature tests into right file
